refactor(web): rename UITestingPanel to UITestingPanelClient and update routes
