+++ apps/web-v2/src/routes/_repo/$owner/$repo/testing-arena/$id.tsx
@@ -37,7 +37,7 @@ import {
   IconTool,
 } from "@tabler/icons-react";
 import dayjs from "@conductor/shared/dates";
-import { UITestingPanel } from "./UITestingPanel";
+import { UITestingPanel } from "./UITestingPanelClient";
 import { parseActivitySteps } from "@/lib/utils/parseActivitySteps";
 import { BranchSelect } from "@/lib/components/BranchSelect";
